requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jsonschema==4.23.0
pytest==8.3.2
httpx==0.27.0
pydantic-settings==2.4.0